fix pointer-type parameters.
In old version, the key of container is verbose output of the given
parameters. Pointer-type parameters are treated as different keys, it
will cause most waste of resources. To solve this, we could use the
original value with `reflect` help. And to avoid parameters dirty, we
use `#` as a division between parameters.